refactor generate_jwt function name, developer application functionality mostly complete
The core functionality of the developer applications is now implemented. Regular user logins still need to be implemented in order for developer applications to be usable. I also need to include endpoints to list developer apps owned by the current user. Regular user logins may require the implementation of a UI login for OAuth logins to work.