add patch for developer endpoint callback URIs and get for details on a user's developer app
I have added a method to replace the existing list of callback URIs with a new one. This may not have been the best way to do this and could cause performance issues or be a really nice way to DoS our database, we should review this. I have also added a way for users to view specific developer applications, this will allow viewing of callback URIs.