Refactor auth flow to use application ID and callback URI, refactor developer login to use new auth flow, add env var for abandon auth internal application, remove unused environment variables
I have refactored the login auth flow to accept a developer application ID and callback URI that are used in combination to redirect the user to the verified callback URI after a successful abandon auth login. The internal developer dashboard/index page utilizes this new auth flow and can act as an example.