Fix Pyright errors
I have gone through and added some additional checks that should have been included and that Pyright caught. This includes some better error handling. I also added a copied error response for the 403 on the /ui/login endpoint. This must be refactored so that the error is raised within the same check for dev_app existing and the callback URIs not being empty as it could be abused to determine valid application IDs, which may be undesirable--better safe than sorry.