create initial AbandonAuth site
This initial site is mostly just a copy of the styling and theming of the current version of the AbandonTech site. This design is highly subject to change and nothing here is final. Mostly here to start transferring the existing Jinja template site to Nuxt, new designs can be implemented later.